refactor: renamed defrasdk package to defradb
