+++ pkg/defradb/network_handler.go
@@ -1,4 +1,4 @@
-package defrasdk
+package defradb
 
 import (
 	"context"
