+++ pkg/defradb/network_handler_test.go
@@ -1,4 +1,4 @@
-package defrasdk
+package defradb
 
 import (
 	"testing"
