+++ pkg/defradb/network_state.go
@@ -1,4 +1,4 @@
-package defrasdk
+package defradb
 
 import (
 	"time"
